EOQ calculator
The software has the goal of helping the logistic division of a company to calculate the EOQ and therefore help them to adjust the quantity of a product based on three parameters: annual cost for holding, annual quantity used and cost for each order. It has a simply interface, make the use of the software possible for any kind of user.